refactor: move onStonePositionChange function into child component
