+++ app/[teamId]/[recordId]/components/stonePositionsSection.tsx
@@ -1,4 +1,4 @@
-import { Coordinate, RecordDetail } from '@/types/model';
+import { RecordDetail } from '@/types/model';
 import React from 'react';
 import { Sheet } from './sheet/sheet';
 
@@ -8,7 +8,6 @@ type Props = {
   selectedEndIndex: number;
   selectedShotIndex: number;
   setSelectedShotIndex: React.Dispatch<React.SetStateAction<number>>;
-  onStonePositionChange: (endIndex: number, shotIndex: number, isFriendStone: boolean, newPosition: Coordinate) => void;
   isEditMode: boolean;
 };
 
@@ -18,9 +17,9 @@ export default function StonePositionsSection({
   selectedEndIndex,
   selectedShotIndex,
   setSelectedShotIndex,
-  onStonePositionChange,
   isEditMode,
 }: Props) {
+
   const selectedShotsData = record.ends_data?.[selectedEndIndex]?.shots ?? [];
   const selectedShotData = selectedShotsData[selectedShotIndex];
 
@@ -39,7 +38,6 @@ export default function StonePositionsSection({
           selectedEndIndex={selectedEndIndex}
           selectedShotIndex={selectedShotIndex}
           setSelectedShotIndex={setSelectedShotIndex}
-          onStonePositionChange={onStonePositionChange}
         />
       </div>
     </section>
