+++ app/[teamId]/[recordId]/recordClient.tsx
@@ -6,7 +6,7 @@ import ScoreBoardSection from './components/scoreBoardSection';
 import MatchDetailsSection from './components/matchDetailsSection';
 import StonePositionsSection from './components/stonePositionsSection';
 import { getRecordDetailsByRecordIdResponse, getTeamDetailsResponse } from '@/types/response';
-import { Coordinate, RecordDetail } from '@/types/model';
+import { RecordDetail } from '@/types/model';
 import { updateRecord } from '@/lib/api/record';
 import { updateRecordRequest } from '@/types/request';
 
@@ -66,42 +66,6 @@ export default function EditableRecordClient({ recordRes, teamRes, recordId }: P
       return { ...prevRecord, is_first: isFirst };
     });
   }
-
-  const handleStonePositionChange = (
-    endIndex: number,
-    shotIndex: number,
-    isFriendStone: boolean,
-    newPosition: Coordinate
-  ) => {
-    setEditedRecord(prevRecord => {
-      const newEndsData = [...prevRecord.ends_data];
-      const targetShot = newEndsData[endIndex].shots[shotIndex];
-      const stoneKey = isFriendStone ? 'friend_stones' : 'enemy_stones';
-      const oldStones = [...targetShot.stones[stoneKey]];
-      const stoneExists = oldStones.some(stone => stone.index === newPosition.index);
-
-      let newStones;
-
-      // Add or move a stone.
-      if (stoneExists) { // 既に存在する石の場合
-        newStones = oldStones.map((stone) => {
-          if (stone.index === newPosition.index) {
-            return newPosition;
-          } else {
-            return stone;
-          }
-        }
-        );
-      } else {
-        newStones = [...oldStones, newPosition];
-      }
-
-      // Update stones in current shot
-      targetShot.stones[stoneKey] = newStones;
-
-      return { ...prevRecord, ends_data: newEndsData };
-    });
-  };
 
   const handleShotsDetailsChange = (
     endIndex: number,
@@ -179,7 +143,6 @@ export default function EditableRecordClient({ recordRes, teamRes, recordId }: P
             selectedShotIndex={selectedShotIndex}
             setSelectedShotIndex={setSelectedShotIndex}
             isEditMode={isEditMode}
-            onStonePositionChange={handleStonePositionChange}
           />
         </div>
       </div>
